refactor: change calculateTripleQuotedStringPadding to use PaddingTracker
Progress toward these issues:
https://github.com/decaffeinate/decaffeinate/issues/557
https://github.com/decaffeinate/decaffeinate/issues/556

This change gets rid of IndexRangeList and the related code in favor of using
PaddingTracker for everything. A lot of code could be removed, but for the other
code in calculateTripleQuotedStringPadding, I tried to keep it similar to the
previous implementation, so the behavior should be the same. This should put the
code in a better position to be reworked to fix the known bugs.